api/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
python-dotenv
pillow